sap-smart-agent/requirements.txt
mcp
httpx
boto3
strands-agents
strands-agents-tools
